blog-core/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
humanize==4.9.0
pytz==2024.1
requests==2.31.0
markdown==3.5.2 